refactor(renderer): extract MobSerializer into _serializer.py
All mobject→wire-dict methods move to MobSerializer; CaptureRenderer
delegates only its public surface (state_refs, state_ref_for,
serialize_mobject, short_id, _state_registry property). Tests that
target serializer internals now use _fresh_serializer() directly instead
of going through the renderer.
